extract optionstostring
